Fix 'rapid-sucessive clicks' nextWord button bug
When clicking or tapping the nextWord button twice in a row
sufficiently fast enough, the first click event is not
finished before the 2nd click event begins. This results in the same
set of letter tiles being generated twice, giving twice as many tiles
as there should be. Fixed by disabling the button until the first event
is complete.